team_jacaranda/coursework_one/pyproject.toml
[tool.poetry]
name = "coursework-one"
version = "0.1.0"
description = ""
authors = ["TeamJacaranda"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
google-api-python-client = "^2.36.0"
psycopg2-binary = "^2.9.9"
logging = "^0.4.9.6"
pathlib = "^1.0.1"
urllib3 = "^1.26.7"
kafka-python = "^2.0.2"
minio = "^7.1.0"
spacy = "^3.5.0"
nltk = "^3.8"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
